Move to subscriber.SingleEmitter
