+++ src/main/java/reactor/core/subscriber/SingleEmitter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package reactor.core.publisher;
+package reactor.core.subscriber;
 
 import reactor.core.flow.Cancellation;
